requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
fpdf
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
transformers
torch